Simplify Mention
