+++ src/index.ts
@@ -5,6 +5,7 @@ export * from './store';
 export * from './components';
 
 export { default as views } from './views';
+export type { NodeViewProps } from './views';
 
 export { setup, opts } from './connect';
 export type { Options } from './connect';
