+++ styles/_mention.scss
@@ -1,0 +1,5 @@
+.mention {
+  background-color: #e0e0e0;
+  border-radius: 5px;
+  padding: 2px 4px;
+}
