+++ styles/index.scss
@@ -14,6 +14,7 @@
 @use 'footnote';
 @use 'figure';
 @use 'codemark';
+@use 'mention';
 
 .hidden {
   display: none;
